separated code between parser and geo, geo passes
